Added check for API backend host existence to redeploy script for new environments.
Added check for API backend host existence to redeploy script for new environments.

+++ devops/deploy/backend/redeployApiBackend.sh
@@ -4,5 +4,9 @@ if [ -z $1 ]; then
   exit -1
 fi;
 instance_id=`aws ec2 describe-instances --filters "Name=tag:EnvironmentName,Values=$1" "Name=instance-state-name,Values=running" | jq -r '.Reservations[0].Instances[0].InstanceId'`
-echo $instance_id
-aws ec2 terminate-instances --instance-ids $instance_id
+if [ $instance_id == "null" ]; then
+  echo "No instance found for environment $1, leaving it alone for now!"
+else
+  echo $instance_id
+  aws ec2 terminate-instances --instance-ids $instance_id
+fi
